Abstracted some common viewset things into a StekViewSet
